pc_client/requirements.txt
PySide2
pyserial
loguru
